perf(deps): optimize dependency features to reduce compile time and binary size
- Optimize tokio features: replace 'full' with specific needed features
  (net, rt, rt-multi-thread, signal, sync, time, macros)
  Removes unused: fs, process, parking_lot, io-std

- Optimize tower features: replace 'full' with specific needed features
  (util, timeout, limit, buffer, make)
  Removes unused: balance, discover, hedge, retry, spawn-ready, load-shed

- Optimize tower-http compression: change from compression-full to compression-gzip
  Removes unused: brotli, deflate, zstd compression formats

- Remove unused axum multipart feature

- Optimize chrono: disable default features, enable only needed ones
  (clock, std, serde)
  Removes: oldtime (deprecated), wasmbind

- Fix clippy::unnecessary_cast in pool_health.rs

Expected impact: 15-25% faster compile times, 10-20% smaller binaries

+++ acton-service/src/pool_health.rs
@@ -72,7 +72,7 @@ pub struct RedisPoolHealth {
 impl RedisPoolHealth {
     /// Create health metrics from a Redis pool
     pub fn from_pool(pool: &deadpool_redis::Pool, config: &crate::config::RedisConfig) -> Self {
-        let max_size = config.max_connections as usize;
+        let max_size = config.max_connections;
         let status = pool.status();
 
         // Pool is available if it's not closed
